Start to move the last of the "global" funcs out of lib.php
These are used in enough places that I'm going to let them remain semiglobal
in helpers but at least this way they will be namespaced, explicitly imported,
and visible to psalm when static checking.
